+++ R/julia-base.R
@@ -24,20 +24,34 @@
 #' plot(as.raster(z), interpolate = FALSE)
 #' @export
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-julia_r <- function(cx = -0.7, cy = 0.27015, size = 400, max_iter = 255) {
+julia_r <- function(c_re = -0.7, c_im = 0.27015, 
+                    x = 0, y = 0, zoom = 1,
+                    width = 400, height = 200, 
+                    max_iter = 255) {
+
+  # Adjust extents
+  dx <- 2
+  dy <- 2
+  if (width > height) {
+    dy <- dy * height / width
+  } else {
+    dx <- dx * width / height
+  }
   
-  N <- size
+  # zoom
+  dx <- dx / zoom
+  dy <- dy / zoom
 
   # Define the (x, y) grid of coordiantes
-  x <- matrix(seq(-2, 2, length.out = N), N, N, byrow = TRUE)
-  y <- matrix(seq(-2, 2, length.out = N), N, N, byrow = FALSE)
+  x <- matrix(seq(-dx + x, dx + x, length.out =  width), height, width, byrow = TRUE)
+  y <- matrix(seq(-dy + y, dy + y, length.out = height), height, width, byrow = FALSE)
   
   # Where is the iteration still valid (i.e. hasn't diverged to infinity)
-  valid <- rep(T, N * N)
+  valid <- rep(T, height * width)
   idx   <- which(valid)
   
   # A matrix to hold the iteration count
-  iter  <- matrix(0L, N, N)
+  iter  <- matrix(0L, height, width)
   
   for (i in seq(max_iter)) {
     # Just calculate on the (x,y) which are still valid
@@ -59,8 +73,8 @@ julia_r <- function(cx = -0.7, cy = 0.27015, size = 400, max_iter = 255) {
     yt <- y[idx]
     
     # Iterative calculation
-    tmp    <- xt * xt - yt * yt + cx 
-    y[idx] <- 2 * xt * yt + cy       
+    tmp    <- xt * xt - yt * yt + c_re 
+    y[idx] <- 2 * xt * yt + c_im       
     x[idx] <- tmp
   }
   
@@ -74,6 +88,8 @@ if (FALSE) {
   #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   # Simple plot
   #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  iter <- julia_r(zoom = 2, x = 0.5)
+  
   cols <- viridisLite::viridis(256, begin = 0.1, end = 0.9)
   cols <- terrain.colors(256) 
   
@@ -81,8 +97,16 @@ if (FALSE) {
   z[] <- cols[z]
   plot(as.raster(z), interpolate = FALSE)
   
-  z <- sqrt(x * x + y * y)
-  z <- round((z / max(z)) ^ gamma * 255) + 1
-  z[] <- cols[z]
-  plot(as.raster(z), interpolate = FALSE)
+  
+  grid::grid.newpage()
+  julia(zoom = 2, x = 0.5, result = 'nara', height = 200) |> grid::grid.raster()
+  
+  
 }
+
+
+
+
+
+
+